Fix crash on fw 9.1.0+
This crash was because on fw 9.1.0+, the number of sessions for pctl was limited to one for the applet. This meant that if another session was open when trying to launch the appley (in this case, change the pin), it would fatal.

Thank you to Behemoth for finding this issue and thanks to WerWolv for telling me about it. Also thanks to Daniel C for testing this fix for me.